+++ src/components/charts/CpuUsage.vue
@@ -1,21 +1,20 @@
 <!-- src/components/charts/CpuUsage.vue -->
 <template>
-  <div class="cpu-usage">
-    <h4>CPU 使用率</h4>
-    <div class="chart-wrapper">
-      <ChartComponent :type="'line'" :data="chartData" :options="chartOptions" />
-    </div>
-  </div>
+  <ChartCard title="CPU 使用率">
+    <ChartComponent type="line" :data="chartData" :options="chartOptions" />
+  </ChartCard>
 </template>
 
 <script>
 import { computed, onMounted } from 'vue'
 import { useTelemetryStore } from '@/stores/telemetry'
+import ChartCard from './ChartCard.vue'
 import ChartComponent from './ChartComponent.vue'
 
 export default {
   name: 'CpuUsage',
   components: {
+    ChartCard,
     ChartComponent
   },
   setup() {
@@ -74,15 +73,3 @@ export default {
   }
 }
 </script>
-
-<style scoped>
-.cpu-usage {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-.chart-wrapper {
-  flex: 1;
-  position: relative;
-}
-</style>